+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Action setvalue/odk-instance-first-load and setvalue/odk-new-repeat actions are not properly added for radiobutton and checkbox questions (in enketo-transformer).
 - Lazy and/or evaluation within function arguments (in openrosa-xpath-evaluator).
 - Action setvalue/odk-new-repeat does not run non-form-control actions before form controls.
+- Static itemsets with radio buttons inside multiple repeat instances do not load correctly.
 
 [5.17.0] - 2020-12-28
 ----------------------